pp/rest: rethrow 401 as 403 on failed authn
The authenticator will throw 401 instead of 403 from
require_authenticated because there are a few endpoints on the Admin API
that do not require authentication but the authorization policy does.

A failure occurs at this point when HTTP Basic AuthN is set in the yaml
config file but the request does not have the authorization header.
Typically, 401 Unauthorized error means the server does not know the
clients identity. 403 Forbidden, however, means authorization failed but
the clients identity is known which is the situation here since user
credentials are validated already. So throwing 401 is misleading.

This commit rethrows that 401 as a 403.